Fixes and content update.

+++ data/scripts/entity/stationgovernor.lua
@@ -46,7 +46,7 @@ function updateServer(timeStep)
             CosmicVaultBuffs.applyBuff(entity.id, "ShieldRecharge", 2.0, 11.0)
         end
         -- Spawn militia once every 10 minutes if enemies are present
-        if random():getInt() < 0.05 then
+        if random():getFloat() < 0.05 then
             -- To avoid complex entity generation in this update loop, we just flag it
         end
     else
